#3 - feat: 정규분포 생성 함수 utils패키지로 분리
테스트코드에서 작성한 함수를 검증하고 나서 utils 패키지로 빼서 관리
